% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case1_z_eu.R
\name{case1_z_eu}
\alias{case1_z_eu}
\title{Euphotic depth function based on Morel and Berthon (1989, Limnol. Oceanogr., Vol 34: 1545-1562) Case I model}
\usage{
case1_z_eu(chl_mean)
}
\arguments{
\item{chl_mean}{A numeric vector of chlorophyll_a surface concentration in milligrams chlorophyl per cubic meter}
}
\value{
euphotic depth in meter
}
\description{
Convert chlorophyll concentration to euphotic depth
}
\references{
Morel A, Berthon J-F (1989) Surface pigments, algal biomass profiles, and potential production of the euphotic layer: Relationships reinvestigated in view of remote-sensing applications. Limnology and Oceanography 34:1545–1562
}
\author{
Chih-Lin Wei <chihlinwei@gmail.com>
}
