% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Bottom_WOA.R
\name{Get_Bottom_WOA}
\alias{Get_Bottom_WOA}
\title{Download near-bottom climatology data from World Ocean Atlas}
\usage{
Get_Bottom_WOA(url, x = NULL, outdir = NULL, plot = FALSE)
}
\arguments{
\item{url}{A character string naming the URL of a CSV file to be downloaded}

\item{x}{Raster* object defining the spatial extent, resolution and projection of climatology data 
to be retrived. If not provide, returning full data set.}

\item{outdir}{output directory to write raster file}

\item{plot}{Logical, whether or not to plot the output raster}
}
\value{
a raster object
}
\description{
Download and near-bottom climatology data from Nationa Oceanographic Data Center (NODC) World Ocean Atlas 2013 or Regional Climatology and then convert to a raster object
}
\details{
This function dowloads the near-bottom climatology from 
\url{http://www.nodc.noaa.gov/OC5/woa13/woa13data.html} or \url{http://www.nodc.noaa.gov/OC5/regional_climate/}.
The global or regional climatology data at standard depths will be download but only the near-bottom
data are retained based on a template raster object.
}
\examples{
# Raster template of Canadian Arctic
data(tmp)

# Download January temperature data as example
Get_Bottom_WOA("https://www.ncei.noaa.gov/data/oceans/woa/WOA18/DATA/temperature/csv/decav/0.25/woa18_decav_t01an04.csv.gz", 
tmp, plot=TRUE) 

#Download monthly near-bottom temperature
urls <- paste("https://www.ncei.noaa.gov/data/oceans/woa/WOA18/DATA/temperature/csv/decav/0.25/woa18_decav_t", 
              c(paste(0, 1:9, sep=""), 10:12), "an04.csv", sep="")

# Raster template of Canadian Arctic
data(tmp)

# Create a fold to save raster file
folder <- "../Arctic_ocean_data/Climatology2" 
dir.create(folder)

library(foreach)
library(doSNOW)

cl<-makeCluster(4) # change 4 to your number of CPU cores
registerDoSNOW(cl) # register the SNOW parallel backend with the foreach package
foreach(i=urls, .packages="OceanData", .export = "urls") \%dopar\% Get_Bottom_WOA(i, tmp, folder)
stopCluster(cl) # stop a SNOW cluster
}
\author{
Chih-Lin Wei <chihlinwei@gmail.com>
}
