% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Interp_WOA.R
\name{Get_IsoValueDepth_WOA}
\alias{Get_IsoValueDepth_WOA}
\title{Download climatology data from World Ocean Atlas and calculate the iso-value depth}
\usage{
Get_IsoValueDepth_WOA(
  url,
  IsoValue = 10,
  x = NULL,
  outdir = NULL,
  plot = FALSE
)
}
\arguments{
\item{url}{A character string naming the URL of a CSV file to be downloaded}

\item{IsoValue}{A numeric values specifying where interpolation is to take place.}

\item{x}{Raster* object defining the spatial extent, resolution and projection of climatology data 
to be retrived. If not provide, returning full data set.}

\item{outdir}{output directory to write raster file}

\item{plot}{Logical, whether or not to plot the output raster}
}
\value{
a raster object
}
\description{
Download and surface climatology data from Nationa Oceanographic Data Center (NODC) World Ocean Atlas 2013 or Regional Climatology, calculate iso-value depth surface and then convert to a raster object
}
\details{
This function dowloads the surface climatology from 
\url{http://www.nodc.noaa.gov/OC5/woa13/woa13data.html} or \url{http://www.nodc.noaa.gov/OC5/regional_climate/}.
The global or regional climatology data at standard depths will be download but only the iso-value depth surface
are retained based on a template raster object.
}
\examples{
# Raster template of Canadian Arctic
data(tmp)

# Download January temperature data and calculate the 10-degree depth surface
Get_IsoValueDepth_WOA("http://data.nodc.noaa.gov/woa/WOA13/DATAv2/temperature/csv/decav/0.25/woa13_decav_t01an04v2.csv.gz", 
IsoValue=10, tmp, plot=TRUE) 

}
\author{
Chih-Lin Wei <chihlinwei@gmail.com>
}
