% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Get_Ocean_Prod.R
\name{Get_Ocean_Prod}
\alias{Get_Ocean_Prod}
\title{Download SeaWiFS or MODIS HDF4 file from Ocean Productivity Web page}
\usage{
Get_Ocean_Prod(url, x = NULL, outdir = NULL, plot = FALSE)
}
\arguments{
\item{url}{A character string naming the URL of a HDF4 file to be downloaded}

\item{x}{Raster* object defining the spatial extent, resolution and projection of climatology data
to be retrived. If not provide, returning full data set.}

\item{outdir}{output directory to write raster file}

\item{plot}{Logical, whether or not to plot the output raster}
}
\value{
a raster object
}
\description{
Download SeaWiFS or MODIS HDF4 file from Ocean Productivity Web page and then convert to a raster object
}
\details{
This function dowloads HDF4 file from Ocean Productivity Web page \url{http://www.science.oregonstate.edu/ocean.productivity/}
This is an R wrapper for the 'gdal_translate' function that is part of the Geospatial Data Abstraction Library (GDAL), which assumes the
user has a working GDAL on their system.
}
\examples{
# Raster template of Canadian Arctic
data(tmp)

# Download August 1998 VGPM as example
url <- "http://orca.science.oregonstate.edu/data/2x4/monthly/vgpm.r2010.s.chl.a.sst/hdf/vgpm.1998182.hdf.gz"
Get_Ocean_Prod(url, x = tmp, plot=TRUE)
}
\author{
Chih-Lin Wei <chihlinwei@gmail.com>
}
\references{
\url{http://www.gdal.org/gdal_translate.html}
}

