% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Surface_WOA.R
\name{Get_Surface_WOA}
\alias{Get_Surface_WOA}
\title{Download surface climatology data from World Ocean Atlas}
\usage{
Get_Surface_WOA(url, x = NULL, outdir = NULL, plot = FALSE)
}
\arguments{
\item{url}{A character string naming the URL of a CSV file to be downloaded}

\item{x}{Raster* object defining the spatial extent, resolution and projection of climatology data 
to be retrived. If not provide, returning full data set.}

\item{outdir}{output directory to write raster file}

\item{plot}{Logical, whether or not to plot the output raster}
}
\value{
a raster object
}
\description{
Download and surface climatology data from Nationa Oceanographic Data Center (NODC) World Ocean Atlas 2013 or Regional Climatology and then convert to a raster object
}
\details{
This function dowloads the surface climatology from 
\url{http://www.nodc.noaa.gov/OC5/woa13/woa13data.html} or \url{http://www.nodc.noaa.gov/OC5/regional_climate/}.
The global or regional climatology data at standard depths will be download but only the surface
data are retained based on a template raster object.
}
\examples{
# Raster template of Canadian Arctic
data(tmp)

# Download January temperature data as example
Get_Surface_WOA("http://data.nodc.noaa.gov/woa/WOA13/DATAv2/temperature/csv/decav/0.25/woa13_decav_t01an04v2.csv.gz", 
tmp, plot=TRUE) 

#Download monthly surface temperature
urls <- paste("http://data.nodc.noaa.gov/woa/WOA13/DATAv2/temperature/csv/decav/0.25/woa13_decav_t", 
              c(paste(0, 1:9, sep=""), 10:12), "an04v2.csv", sep="")

# Raster template of Canadian Arctic
data(tmp)

# Create a fold to save raster file
folder <- "../Arctic_ocean_data/Climatology2" 
dir.create(folder)

library(foreach)
library(doSNOW)

cl<-makeCluster(4) # change 4 to your number of CPU cores
registerDoSNOW(cl) # register the SNOW parallel backend with the foreach package
foreach(i=urls, .packages="OceanData", .export = "urls") \%dopar\% Get_Surface_WOA(i, tmp, folder)
stopCluster(cl) # stop a SNOW cluster
}
\author{
Chih-Lin Wei <chihlinwei@gmail.com>
}
