% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loc.R
\docType{data}
\name{loc}
\alias{loc}
\title{Megabenthos sampling location in the Canadian Arctic}
\format{SpatialPointsDataFrame}
\description{
A SpatialPointsDataFrame showing Megabenthos sampling locations in Sinusoidal Eaqual area projection.
}
\author{
Chih-Lin Wei <chihlinwei@gmail.com>
}
\keyword{datasets}

