% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lutz_p_flux.R
\name{prd}
\alias{prd}
\alias{rld}
\alias{prr}
\alias{p.ratio}
\alias{lutz_p_flux}
\title{Export POC flux function based on Lutz et al. (2007, J. Geophys. Res., Vol 112, C10011)}
\usage{
prd(npp_svi)

rld(npp_svi)

prr(npp_svi)

p.ratio(prd, z_ex, rld, prr)

lutz_p_flux(npp_mean, npp_svi, z_ex)
}
\arguments{
\item{npp_svi}{A numeric vector of Seasonal variation index of production (SVI; annual standard deviation divided by average)}

\item{z_ex}{A numeric vector of eport depth = water depth (z) - euphotic depth (z_eu)}

\item{npp_mean}{A numeric vector of nnual mean surface production}
}
\value{
Export POC flux in the same unit as input primary production
}
\description{
Convert mean and seasonal index of primary production and export depth to export POC flux
}
\section{Functions}{
\itemize{
\item \code{prd()}: the labile fraction of export

\item \code{rld()}: remineralization length scale

\item \code{prr()}: rapidly sinking fraction of export

\item \code{p.ratio()}: Annual Particulate Organic Carbon Flux to Depth Normalized to Overlying Production

}}
\examples{
# Seasonal variation index of NPP (SD/mean)
npp_svi <- 0.2789988

# Mean NPP (mgC/m2/day)
npp_mean <- 294.2888

# Mean chlorophyll concentrations (mg/m3)
chl_mean <- 0.123846

# Euphotic depth (m)
z_eu <- case1_z_eu(chl_mean)

# Water depth (m)
z <- 3115.12

# Export depth (m)
z_ex <- z-z_eu

# Lutz et al. (2007) POC flux (mgC/m2/day)
lutz_p_flux(npp_mean, npp_svi, z_ex)
}
\references{
Lutz MJ, Caldeira K, Dunbar RB, Behrenfeld MJ (2007) Seasonal rhythms of net primary production and particulate organic carbon flux to depth describe the efficiency of biological pump in the global ocean. Journal of Geophysical Research: Oceans 112:C10011
}
\author{
Chih-Lin Wei <chihlinwei@gmail.com>
}
