% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Year_Day.R
\name{Year_Day}
\alias{Year_Day}
\title{Regularly spaced year plus day of year}
\usage{
Year_Day(...)
}
\arguments{
\item{...}{Arguments passed to \link[timeDate]{timeSequence}}

\item{from}{starting date, required.}

\item{to}{end date, optional. If supplied, to must be after (later than) from.}

\item{by}{a character string, containing one of "sec", "min", "hour", "day", "week", "month" or "year". 
This can optionally be preceded by an integer and a space, or followed by "s".
character string "quarter" that is equivalent to "3 months".
A number, taken to be in seconds.
A object of class 'difftime'}
}
\value{
a numeric vector of year plus day of year
}
\description{
Regularly spaced year plus day of year. This year day format is commonly used in NASS Ocean Color product
}
\details{
This function's Arguments follow \link[timeDate]{timeSequence}
}
\examples{
Year_Day("1998-01-01", "1999-12-01", by="month")
LazyData: true
}
\author{
Chih-Lin Wei <chihlinwei@gmail.com>
}
